Added column types checks
