+++ eda/src/autogluon/eda/analysis/__init__.py
@@ -1,2 +1,3 @@
-from .univariate import *
 from .composite import *
+from .missing import *
+from .univariate import *
